Refactor instances state check
